integ sa deped

+++ database/migrations/2026_04_08_141355_add_split_names_to_users_table.php
@@ -57,7 +57,9 @@ return new class extends Migration
         }
 
         Schema::table('users', function (Blueprint $table) {
-            $table->dropColumn('name');
+            if (Schema::hasColumn('users', 'name')) {
+                $table->dropColumn('name');
+            }
         });
     }
